% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{msgpack_format}
\alias{msgpack_format}
\title{Format data for msgpack}
\usage{
msgpack_format(x)
}
\arguments{
\item{x}{An r object.}
}
\value{
A formatted R object to use as input to msgpack_pack.
}
\description{
A helper function to format R data for input to msgpack
}
