% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{msgpack_pack}
\alias{msgpack_pack}
\title{Msgpack Pack}
\usage{
msgpack_pack(...)
}
\arguments{
\item{...}{Any R objects that have corresponding msgpack types.}
}
\value{
A raw vector containing the message.
}
\description{
Serialize any number of objects into a single message.
}
\seealso{
See test.r for examples in the package directory.
}
