% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{msgpack_simplify}
\alias{msgpack_simplify}
\title{Simplify msgpack}
\usage{
msgpack_simplify(x)
}
\arguments{
\item{x}{Return object from msgpack_unpack.}
}
\value{
A simplified return object from msgpack_unpack.  E.g., arrays of all the same type are concatenated into an atomic vector.
}
\description{
A helper function for simplifying a msgpack return object
}
